chore(gui): projects page minor improved
